requirements.txt
biopython>=1.79
pandas>=1.3.0
plotly>=5.0.0
kaleido>=0.1.0